requirements.txt
pydantic~=2.10.4
openai~=1.58.1
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
datasets~=3.2.0
fastapi>=0.68.0

html2text~=2024.2.26
gymnasium~=1.0.0
pillow~=10.4.0
browsergym~=0.13.3
uvicorn>=0.15.0
unidiff~=0.7.5
browser-use~=0.1.40
googlesearch-python~=1.3.0

aiofiles~=24.1.0
pydantic_core~=2.27.2
colorama~=0.4.6
playwright~=1.49.1
websockets>=10.0
